+++ tests/cli_tests/delete.rs
