+++ tests/cli_tests/diff.rs
