+++ tests/cli_tests/unix/diff.rs
